refactor: 행사 업로드 실패 로직 수정

+++ event/src/main/java/life/eventory/event/service/impl/EventServiceImpl.java
@@ -60,7 +60,9 @@ public class EventServiceImpl implements EventService {
             );
         }  catch (Exception e) {
             log.error(e.getMessage());
-            communicationService.deletePoster(imageId);
+            if(hasImage) {
+                communicationService.deletePoster(imageId);
+            }
             throw new IllegalStateException(e.getMessage());
         }
     }
